+++ src/test/kotlin/org/jetbrains/intellij/platform/gradle/tasks/PrintBundledModulesTaskTest.kt
@@ -1,0 +1,22 @@
+// Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
+
+package org.jetbrains.intellij.platform.gradle.tasks
+
+import org.jetbrains.intellij.platform.gradle.Constants.Tasks
+import org.jetbrains.intellij.platform.gradle.IntelliJPluginTestBase
+import org.jetbrains.intellij.platform.gradle.assertContains
+import kotlin.test.Test
+
+class PrintBundledModulesTaskTest : IntelliJPluginTestBase() {
+
+    @Test
+    fun `print bundled modules`() {
+        build(Tasks.PRINT_BUNDLED_MODULES) {
+            assertContains("> Task :${Tasks.PRINT_BUNDLED_MODULES}", output)
+            assertContains("Bundled modules for IntelliJ IDEA", output)
+            assertContains("intellij.platform.coverage", output)
+            assertContains("intellij.platform.coverage.agent", output)
+            assertContains("intellij.platform.vcs.impl", output)
+        }
+    }
+}